Refactored Trades type into types folder
