+++ src/client/src/types/index.ts
@@ -21,6 +21,7 @@ export { ServiceConst } from './serviceConst'
 export { ServiceStatus } from './serviceStatus'
 export { ServiceInstanceStatus } from './serviceInstanceStatus'
 export { Trade } from './trade'
+export { Trades } from './trades'
 export { TradeStatus } from './tradeStatus'
 export { TradesUpdate } from './tradesUpdate'
 export { UpdateType } from './updateType'
