+++ src/client/src/types/trades.ts
@@ -1,0 +1,5 @@
+import { Trade } from './trade'
+
+export interface Trades {
+  [tradeId: number]: Trade
+}
